@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.19.5 <3.0.0'
 
 dependencies:
   flutter:
@@ -25,6 +25,14 @@ dependencies:
   flutter_svg: ^2.0.4
   cloud_firestore: ^4.4.5
   flutter_mapbox_autocomplete: ^2.0.0
+  http: ^0.13.5
+  mapbox_api: ^1.0.0
+  firebase_core: ^2.10.0
+  
+
+  
+  
+  
   
    
 
