@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-//import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:plugyourev/login_page.dart';
 import 'package:plugyourev/register.dart';
 import 'package:plugyourev/splash_screen.dart';
@@ -7,7 +7,7 @@ import 'package:plugyourev/profile_page.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  //await Firebase.initializeApp();
+  await Firebase.initializeApp();
   runApp(MaterialApp(
     debugShowCheckedModeBanner: false,
     home: SplashScreen(),
